backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.0.350
openai==1.3.5
faiss-cpu==1.7.4
python-dotenv==1.0.0
pypdf==3.17.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
flask==2.3.3
flask-cors==4.0.0
pymongo==4.6.1 